I'm opening this as a draft in case folks want to weigh in on the changes. I've still got a few changes I want to make over the next couple weeks so this isn't quite ready yet. This PR includes the commit from #120, but that can be merged independently so it doesn't block #119.

The PR may be easier to read if you read the commits independently. I forgot @samth also worked on removing `run-server` until I looked at the list of open PRs just now, but I took that a little further by also improving the server loop in a couple ways so I think this PR supersedes #106.

Tested against the `/plaintext` endpoint of [this app](https://github.com/Bogdanp/FrameworkBenchmarks/blob/87870c5c1d746e8c253fdd3bf03e18eec4dafc69/frameworks/Racket/racket/app.rkt). Master branch:

```
Requests      [total, rate, throughput]         30000, 1000.04, 1000.03
Duration      [total, attack, wait]             29.999s, 29.999s, 286.417µs
Latencies     [min, mean, 50, 90, 95, 99, max]  229.205µs, 318.265µs, 293.078µs, 367.174µs, 393.675µs, 606.119µs, 14.015ms
Bytes In      [total, mean]                     390000, 13.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:30000  
Error Set:
```

(note that this app uses its own [dispatcher](https://github.com/Bogdanp/FrameworkBenchmarks/blob/87870c5c1d746e8c253fdd3bf03e18eec4dafc69/frameworks/Racket/racket/app.rkt#L217-L218), not `dispatch/servlet`, which is why these numbers are better than the baseline numbers in commit 4506e7290b30d4387f6c63a0f8c95b97782fecab -- `dispatch/servlet` needs to do a lot of work that applications that don't use continuations don't benefit from)

This branch:

```
Requests      [total, rate, throughput]         30000, 1000.04, 1000.03
Duration      [total, attack, wait]             29.999s, 29.999s, 262.316µs
Latencies     [min, mean, 50, 90, 95, 99, max]  197.619µs, 253.639µs, 248.664µs, 280.93µs, 297.689µs, 344.797µs, 929.773µs
Bytes In      [total, mean]                     390000, 13.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:30000
Error Set:
```

For comparison, here's how Node 17 running express does on the same benchmark on my machine:

```
Requests      [total, rate, throughput]         30000, 1000.03, 1000.03
Duration      [total, attack, wait]             29.999s, 29.999s, 182.925µs
Latencies     [min, mean, 50, 90, 95, 99, max]  136.996µs, 190.037µs, 185.03µs, 208.014µs, 219.565µs, 303.124µs, 1.406ms
Bytes In      [total, mean]                     390000, 13.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:30000
Error Set:
```