@samth 2022-06-15
Does removing `run-server` mean it no longer depends on compatibility-lib?