@Bogdanp 2022-06-16
> Does removing run-server mean it no longer depends on compatibility-lib?

Yup, looks like it does.