@LiberalArtist 2022-06-16
Re replacing `run-server`:

It's a pathological case, but mechanisms like `(kill-thread (current-thread))` and `(custodian-shutdown-all (current-custodian))` escape without running winders, e.g.:
```racket
#lang racket
(dynamic-wind
 void
 (λ ()
   (kill-thread (current-thread)))
 (λ ()
   (println 'this-never-prints)))
```
I think `run-server`'s cleanup thread handles this by `sync`ing on the handler thread (without the vestigial timeout, it might be more efficient to go via `thread-dead-evt`), but I think this replacement would not call `custodian-shutdown-all` on the custodian that owns the ports.

This sort of subtlety is why I suggested in <https://github.com/racket/web-server/pull/106#issuecomment-882132036>:
> If it seems better to copy–paste for now, I'd add a comment here and maybe even in `mzlib/thread` in the hope of prompting someone to tackle the refactoring in the future.

More broadly, this is reminding me of mailing-list discussion in https://groups.google.com/g/racket-users/c/AAeEp_llxaY/m/uRviksPGAgAJ and https://groups.google.com/g/racket-dev/c/PyetqHSjtAA/m/slBdazdqAwAJ about the subtleties of delimited control and the Concurrent ML subsystem and the limitations of finalization with `dynamic-wind`. (I'd recently recommended these threads to people adapting the [Shepherd](https://www.gnu.org/software/shepherd/manual/shepherd.html) (a daemon-managing daemon) to use Guile's Fibers.) In particular, @lexi-lambda pointed out the technical report [“Algebraic Effect
Handlers with Resources and Deep Finalization”](https://www.microsoft.com/en-us/research/uploads/prod/2018/04/resource-v1.pdf) (which I want to finish reading!), which says in §8.7 "Dynamic Wind" (internal citations omitted):

> The Scheme language always supported delimited continuations and
> they have also struggled with initialization- and finalization
> for continuations that resumed more than once. The
> `unwind-protect` in Scheme is like a `finally` clause, while
> `dynamic-wind` is like `initially`/finally with a pre- and
> postlude. [Sitaram](https://web.archive.org/web/20200223020813/http://www.ccs.neu.edu/~dorai/uwcallcc/uwcallcc.html) describes how the standard dynamic-wind is not
> good enough in general:
>
>> While this may seem like a natural extension of the first-order
>> `unwind-protect` to a higher-order control scenario, it does not
>> tackle the pragmatic need that `unwind-protect` addresses,
>> namely, the need to ensure that a kind of “clean-up” happens only
>> for those jumps that significantly exit the block, and not for
>> those that are a minor excursion.  The crux is identifying which
>> of these two categories a jump falls into.
>
> Interestingly, this is exactly what is addressed by algebraic
> effects where “significant exit”s are operations that do not
> resume, while “minor excursions” are regular operations that
> resume with a result.

I'll try to take a look at other aspects of this PR later.