@Bogdanp 2022-06-16
The `(custodian-shutdown-all (current-custodian))` case is covered by the custodian that wraps the handler, but you're right about the `(kill-thread (current-thread))` case. I don't think the server claims that handler threads are kill-safe, but it's easy enough to bring back the cleanup thread so I'll just do that. An alternative I've considered is keeping track of the handler threads in a list inside the loop and syncing on them to remove them from the set when done. That would get rid of the need for both `dynamic-wind` and a cleanup thread, but would result in more work per loop iteration so it's probably not worth it.

> I think this replacement would not call custodian-shutdown-all on the custodian that owns the ports.

I'm not sure I understand what you mean here. The custodian that gets shut down is the one that wraps both the thread and the TCP ports.