@Bogdanp 2022-06-22
Here's how we're doing so far:

* [master branch](https://www.techempower.com/benchmarks/#section=test&shareid=85490882-b757-4fa0-9340-5babbb12ffc6&hw=ph&test=composite&a=2) (88,090RPS on the plaintext benchmark)
* [latest commit on this branch](https://www.techempower.com/benchmarks/#section=test&shareid=67e37271-e872-44a8-96f1-557e38b7fa06&hw=ph&test=json&a=2) (151,709RPS on the plaintext benchmark)

For comparison, here's the plaintext benchmark on [express.js](https://www.techempower.com/benchmarks/#section=test&shareid=ec422152-d3bc-46b0-9a04-00b46a871f34&hw=ph&test=plaintext&a=2) on the same machine. The test machine in this case is a 24-core AMD server.