@Bogdanp 2022-06-25
I'll wrap this up tomorrow morning and then we can merge. There are more improvements that could be made, but we can leave those for another time so this doesn't get too large and because I'm not sure when I'll have time to do them.